+++ src/main/java/net/minecraft/block/BlockFurnace.java
@@ -67,9 +67,9 @@ public class BlockFurnace extends BlockContainer {
     public void randomDisplayTick(World worldIn, BlockPos pos, IBlockState state, Random rand) {
         if (this.isBurning) {
             EnumFacing enumfacing = (EnumFacing) state.getValue(FACING);
-            double d0 = (double) pos.getX() + 0.5D;
-            double d1 = (double) pos.getY() + rand.nextDouble() * 6.0D / 16.0D;
-            double d2 = (double) pos.getZ() + 0.5D;
+            double d0 = pos.getX() + 0.5D;
+            double d1 = pos.getY() + rand.nextDouble() * 6.0D / 16.0D;
+            double d2 = pos.getZ() + 0.5D;
             double d3 = 0.52D;
             double d4 = rand.nextDouble() * 0.6D - 0.3D;
