+++ src/main/java/net/minecraft/block/BlockLeaves.java
@@ -153,9 +153,9 @@ public abstract class BlockLeaves extends BlockLeavesBase {
 
     public void randomDisplayTick(World worldIn, BlockPos pos, IBlockState state, Random rand) {
         if (worldIn.isRainingAt(pos.up()) && !World.doesBlockHaveSolidTopSurface(worldIn, pos.down()) && rand.nextInt(15) == 1) {
-            double d0 = (double) ((float) pos.getX() + rand.nextFloat());
-            double d1 = (double) pos.getY() - 0.05D;
-            double d2 = (double) ((float) pos.getZ() + rand.nextFloat());
+            double d0 = (pos.getX() + rand.nextFloat());
+            double d1 = pos.getY() - 0.05D;
+            double d2 = (pos.getZ() + rand.nextFloat());
             worldIn.spawnParticle(EnumParticleTypes.DRIP_WATER, d0, d1, d2, 0.0D, 0.0D, 0.0D, new int[0]);
         }
     }
